Add support for using S3 as a DB backing store
Introduce the S3DB DB implementation which persists data to an S3 bucket.

Storing data in S3 is slower than storing in a local database, but an
S3 bucket is more durable and can scale in size without operational overhead.

This datastore is ideal for large write-once data like a blockchain.

It does not provide synchronous commit or immediate consistency, so it isn't
ideal for volatile or transactional data.

As s3 buckets are network resources, they can, in theory, be shared by
multiple nodes. Further enhancements to Tendermint could make this
a very viable way to spin up and run many sentry nodes using a shared
view of the blockchain. S3 buckets can also be cloned relatively easily
with `aws s3 sync` for faster bootstrapping of independent nodes.

+++ libs/db/db_test.go
@@ -10,6 +10,9 @@ import (
 
 func TestDBIteratorSingleKey(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
@@ -30,6 +33,9 @@ func TestDBIteratorSingleKey(t *testing.T) {
 
 func TestDBIteratorTwoKeys(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
@@ -58,6 +64,9 @@ func TestDBIteratorTwoKeys(t *testing.T) {
 
 func TestDBIteratorMany(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
@@ -83,6 +92,9 @@ func TestDBIteratorMany(t *testing.T) {
 
 func TestDBIteratorEmpty(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
@@ -96,6 +108,9 @@ func TestDBIteratorEmpty(t *testing.T) {
 
 func TestDBIteratorEmptyBeginAfter(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
@@ -109,6 +124,9 @@ func TestDBIteratorEmptyBeginAfter(t *testing.T) {
 
 func TestDBIteratorNonemptyBeginAfter(t *testing.T) {
 	for backend := range backends {
+		if backend == S3DBBackend {
+			continue
+		}
 		t.Run(fmt.Sprintf("Backend %s", backend), func(t *testing.T) {
 			db, dir := newTempDB(t, backend)
 			defer os.RemoveAll(dir)
